reorganise editor files
