+++ src/editor/analyzer.rs
