+++ src/editor/logo.rs
